# public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Server",
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:9501",
            "description": "Local server"
        },
        {
            "url": "https://catchupai.on-forge.com",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/v1/custom-prompts/preview": {
            "post": {
                "tags": [
                    "CustomPrompts"
                ],
                "summary": "Run a prompt against one video and return the summary preview",
                "operationId": "api.v1.custom-prompts.preview.store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "media_id",
                                    "content"
                                ],
                                "properties": {
                                    "media_id": {
                                        "type": "string",
                                        "example": "01JCXYZ123456789ABCDEFGHIJ"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "請以學習筆記的風格整理重點。"
                                    },
                                    "model_id": {
                                        "description": "Falls back to the system default when omitted, or when the plan does not allow it.",
                                        "type": "string",
                                        "example": "01k9v7m2q8n4r6t0w3y5z7b1c9",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Preview generated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "short_summary": {
                                            "type": "string",
                                            "example": "一句話總結。"
                                        },
                                        "long_summary": {
                                            "properties": {
                                                "content": {
                                                    "type": "string",
                                                    "example": "完整的長摘要。"
                                                },
                                                "key_points": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "keywords": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "responses": {
...
            "401": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Unauthenticated."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "422": {
                "description": "Invalid request parameters",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "messages": {
                                    "description": "驗證失敗的實際回應。\n\n`App\\Exceptions\\InvalidRequestException::render()` 回的是 422 加上 `messages`\n這個鍵（不是 400、也不是 `errors`）——`$statusCode` 與 `render()` 兩處都寫死了。\n既有端點多半沿用 Http400，那份與執行期行為對不上；新端點請用這一份。",
                                    "type": "object",
                                    "example": {
                                        "redirect": [
                                            "導轉網址為必填。"
                                        ]
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
...
    },
...
}